+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraSettingsTests.swift
@@ -106,7 +106,7 @@ private final class TestMediaSettingsAVWrapper: FLTCamMediaSettingsAVWrapper {
   }
 
   override func recommendedVideoSettingsForAssetWriter(
-    withFileType fileType: AVFileType, for output: FLTCaptureVideoDataOutput
+    withFileType fileType: AVFileType, for output: CaptureVideoDataOutput
   ) -> [String: Any]? {
     return [:]
   }