+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation_objc/include/camera_avfoundation/camera_avfoundation.h
@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import "FLTCamConfiguration.h"
-#import "FLTCameraDeviceDiscovering.h"
 #import "FLTCameraPermissionManager.h"
 #import "FLTCaptureDevice.h"
 #import "QueueUtils.h"