+++ packages/go_router_builder/lib/src/type_helpers.dart
@@ -224,10 +224,11 @@ String compareField(
 }
 
 /// Gets the name of the `const` map generated to help encode [Enum] types.
-String enumMapName(InterfaceType type) => '_\$${type.element.name}EnumMap';
+String enumMapName(InterfaceType type) =>
+    '_\$${type.element.displayName}EnumMap';
 
 /// Gets the name of the `const` map generated to help encode [Json] types.
-String jsonMapName(InterfaceType type) => type.element.name;
+String jsonMapName(InterfaceType type) => type.element.displayName;
 
 String _stateValueAccess(
   FormalParameterElement element,
@@ -302,8 +303,10 @@ class _TypeHelperBigInt extends _TypeHelperWithHelper {
       );
 
   @override
-  bool _matchesType(DartType type) =>
-      const TypeChecker.fromRuntime(BigInt).isAssignableFromType(type);
+  bool _matchesType(DartType type) => const TypeChecker.typeNamed(
+    BigInt,
+    inSdk: true,
+  ).isAssignableFromType(type);
 }
 
 class _TypeHelperBool extends _TypeHelperWithHelper {
@@ -342,8 +345,10 @@ class _TypeHelperDateTime extends _TypeHelperWithHelper {
       );
 
   @override
-  bool _matchesType(DartType type) =>
-      const TypeChecker.fromRuntime(DateTime).isAssignableFromType(type);
+  bool _matchesType(DartType type) => const TypeChecker.typeNamed(
+    DateTime,
+    inSdk: true,
+  ).isAssignableFromType(type);
 }
 
 class _TypeHelperDouble extends _TypeHelperWithHelper {
@@ -472,14 +477,17 @@ class _TypeHelperExtensionType extends _TypeHelper {
         representationType.isDartCoreNum ||
         representationType.isDartCoreBool ||
         representationType.isEnum ||
-        const TypeChecker.fromRuntime(
+        const TypeChecker.typeNamed(
           BigInt,
+          inSdk: true,
         ).isAssignableFromType(representationType) ||
-        const TypeChecker.fromRuntime(
+        const TypeChecker.typeNamed(
           DateTime,
+          inSdk: true,
         ).isAssignableFromType(representationType) ||
-        const TypeChecker.fromRuntime(
+        const TypeChecker.typeNamed(
           Uri,
+          inSdk: true,
         ).isAssignableFromType(representationType);
   }
 }
@@ -566,7 +574,7 @@ class _TypeHelperUri extends _TypeHelperWithHelper {
 
   @override
   bool _matchesType(DartType type) =>
-      const TypeChecker.fromRuntime(Uri).isAssignableFromType(type);
+      const TypeChecker.typeNamed(Uri, inSdk: true).isAssignableFromType(type);
 }
 
 class _TypeHelperIterable extends _TypeHelperWithHelper {
@@ -606,16 +614,18 @@ class _TypeHelperIterable extends _TypeHelperWithHelper {
       // get correct type for iterable
       String iterableCaster = '';
       String fallBack = '';
-      if (const TypeChecker.fromRuntime(
+      if (const TypeChecker.typeNamed(
         List,
+        inSdk: true,
       ).isAssignableFromType(parameterElement.type)) {
         iterableCaster += '.toList()';
         if (!parameterElement.type.isNullableType &&
             !parameterElement.hasDefaultValue) {
           fallBack = '?? const []';
         }
-      } else if (const TypeChecker.fromRuntime(
+      } else if (const TypeChecker.typeNamed(
         Set,
+        inSdk: true,
       ).isAssignableFromType(parameterElement.type)) {
         iterableCaster += '.toSet()';
         if (!parameterElement.type.isNullableType &&
@@ -656,8 +666,10 @@ $fieldName$nullAwareAccess.map((e) => e.toString()).toList()''';
   }
 
   @override
-  bool _matchesType(DartType type) =>
-      const TypeChecker.fromRuntime(Iterable).isAssignableFromType(type);
+  bool _matchesType(DartType type) => const TypeChecker.typeNamed(
+    Iterable,
+    inSdk: true,
+  ).isAssignableFromType(type);
 
   @override
   String _compare(String value1, String value2) =>
@@ -783,10 +795,12 @@ class _TypeHelperJson extends _TypeHelperWithHelper {
     }
 
     final FunctionType functionType = secondParam.type as FunctionType;
-    if (functionType.parameters.length != 1 ||
+    // ignore: experimental_member_use
+    if (functionType.formalParameters.length != 1 ||
         functionType.returnType.getDisplayString() !=
-            type.element.typeParameters.first.getDisplayString() ||
-        functionType.parameters[0].type.getDisplayString() != 'Object?') {
+            type.element.typeParameters.first.displayName ||
+        // ignore: experimental_member_use
+        functionType.formalParameters[0].type.getDisplayString() != 'Object?') {
       throw InvalidGenerationSourceError(
         'The parameter type '
         '`${type.getDisplayString(withNullability: false)}` not have a supported fromJson definition.',